exericse_api/requirements.txt
python-dotenv
requests
firebase-admin